+++ app/src/main/java/org/linphone/activities/voip/viewmodels/CallsViewModel.kt
@@ -24,7 +24,6 @@ import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import org.linphone.LinphoneApplication.Companion.coreContext
-import org.linphone.LinphoneApplication.Companion.corePreferences
 import org.linphone.R
 import org.linphone.activities.voip.data.CallData
 import org.linphone.core.*