+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Group changes to describe their impact on the project, as follows:
 - Confirmation dialog before removing account
 - Attended transfer instead of blind transfer if there is more than 1 call
 - Added hidden setting to disable video completely
+- Added hidden setting to prevent adding / editing / removing native contacts
 
 ### Changed
 - Account EXPIRES is now set to 1 month instead of 1 year for sip.linphone.org accounts